client for user, backend, payment

+++ backend/src/main/java/com/muei/Online/Food/Ordering/model/Order.java
@@ -34,5 +34,5 @@ public class Order {
     @OneToMany
     private List<OrderItem> items;
     private int totalItem;
-    private int totalPrice;
+    private Long totalPrice;
 }
